+++ README.md
@@ -19,11 +19,11 @@ Para visualizar a documentação de cada módulo, utilize os links abaixo:
 - [Guia rápido de atualização](https://github.com/oititec/liveness-angular-example/blob/main/src/app/update-guide/README.md)
 
 
-### [FaceTec SDK: 10.0.42](https://github.com/oititec/liveness-angular-example/blob/main/src/assets/10.0.42)
+### [FaceTec SDK: 10.0.42](https://github.com/oititec/liveness-angular-example/blob/main/src/assets/core-sdk-v10)
 
 _data: 29/04/2026_
 
-- Nova implementação com o liveness 3D da Facetec versão 10
+- Nova implementação de liveness 3D com o SDK web da Facetec versão 10.0.42
 
 ### [Iproov SDK: 6.0.2](https://github.com/oititec/liveness-angular-example/blob/main/src/app/iproov/README.md)
