+++ src/assets/core-sdk-v10/core-sdk-optional/FaceTecStrings.fr.js
