+++ src/assets/core-sdk-v10/core-sdk-optional/FaceTecStrings.pt-br.js
