+++ src/assets/core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecSDK.js
