+++ src/assets/core-sdk-v10/core-sdk/FaceTec_images/FaceTec_document.png
