+++ src/assets/facetec-v10/utilities/DeveloperStatusMessages.ts
@@ -1,5 +1,5 @@
-import { FaceTecInitializationError, FaceTecSessionRequestProcessorCallback, FaceTecSessionStatus } from "../../10.0.42/core-sdk/FaceTecSDK.js/FaceTecPublicApi";
-import { FaceTecSDK } from "../../10.0.42/core-sdk/FaceTecSDK.js/FaceTecSDK";
+import { FaceTecInitializationError, FaceTecSessionRequestProcessorCallback, FaceTecSessionStatus } from "../../core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecPublicApi";
+import { FaceTecSDK } from "../../core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecSDK";
 import { FaceTecStatusEnumFriendlyText } from "./FaceTecStatusEnumFriendlyText";
 
 export class DeveloperStatusMessages {
